Add prepare-release workflow
Note that this changes to the changelog format as specified by the
update-changelog action in realm/ci-actions - while mostly compatible
the version headers no longer start with the "Release" keyword. The
presence of that keyword confuses the action - so the next release
will have to be created with a manually specified version number.

The current realm studio release process keeps changes for the upcoming release
in a separate file (RELEASENOTES.md), and copies its content on top of the
changelog, before resetting it from a template.
The action used by other components keeps the release notes directly in
CHANGELOG.md, and just fills in the version. The format is slightly different
- but close enough that apart from above mentioned first manual
version numbering all tooling should continue working. As other
components use the shared action moving to that is more sensible than
maintaining a custom implementation.